+++ R/JAB_plot.R
@@ -1,14 +1,6 @@
 #' Plots JAB as a function of the p-value
 #'
-#' @param n Sample size
-#' @param BF Bayes factor you would like to match. 1 to avoid the Lindley Paradox, 3 to achieve moderate evidence and 10 to achieve strong evidence.
-#' @param method Used for the choice of 'b'. Currently one of:
-#' \itemize{
-#'   \item "JAB": this choice of b produces Jeffery's approximate BF
-#'   \item "min": uses the minimal training sample for the prior (Gu et al., '17)
-#'   \item "robust": a robust version of "min" that prevents too small b (O'Hagan, '95)
-#'   \item "balanced": this choice of b balances the type I and type II errors
-#' }
+#' @inheritParams alphaN
 #'
 #' @return Prints a plot.
 #' @export
@@ -17,27 +9,20 @@
 #' # Plot JAB as function of the p-value for a sample size of 2000
 #' JAB_plot(2000)
 #' @importFrom graphics abline axis points
-JAB_plot <- function(n, BF=1, method="JAB"){
+JAB_plot <- function(n, BF=1, method="JAB", upper = 1){
 
-  alpha <- alphaN(n = n, BF = BF, method = method)
+  alpha <- alphaN(n = n, BF = BF, method = method, upper = upper)
 
   indicated <- alpha
-  lindley <- alphaN(n = n, BF = 1, method = method)
-  moderate <- alphaN(n = n, BF = 3, method = method)
-  strong <- alphaN(n = n, BF = 10, method = method)
+  lindley <- alphaN(n = n, BF = 1, method = method, upper = upper)
+  moderate <- alphaN(n = n, BF = 3, method = method, upper = upper)
+  strong <- alphaN(n = n, BF = 10, method = method, upper = upper)
 
   ### and make a sequence of t-stats to get p-values
   ts <- seq(1.2, 5, length.out=150)
   ps <- 1-pchisq(ts^2, 1)
-  ## Third, loop to get a Bayes factor every p-value
-
-  bf <- numeric(length(ts))
-  i <- 0
-
-  for(t in ts){
-    i <- i+1
-    bf[i] <- JABt(n = n, t = t, method = method)
-  }
+  ## Third, get a Bayes factor for every p-value
+  bf <- JABt(n = n, t = ts, method = method, upper = upper)
 
   plot(ps, bf, type="l", lty=1, lwd=2, log = "y",
        xlab = bquote(italic("p")*"-value"), ylab = "Bayes factor",